Extract Footprints functionality
